+++ services/ocr/v1/model/RecognizeCambodianIdCardRequest.ts
@@ -1,0 +1,12 @@
+import { CambodianIdCardRequestBody } from './CambodianIdCardRequestBody';
+
+
+export class RecognizeCambodianIdCardRequest {
+    public body?: CambodianIdCardRequestBody;
+    public constructor() { 
+    }
+    public withBody(body: CambodianIdCardRequestBody): RecognizeCambodianIdCardRequest {
+        this['body'] = body;
+        return this;
+    }
+}