+++ services/ocr/v1/model/RecognizeExitEntryPermitRequest.ts
@@ -1,0 +1,12 @@
+import { ExitEntryPermitRequestBody } from './ExitEntryPermitRequestBody';
+
+
+export class RecognizeExitEntryPermitRequest {
+    public body?: ExitEntryPermitRequestBody;
+    public constructor() { 
+    }
+    public withBody(body: ExitEntryPermitRequestBody): RecognizeExitEntryPermitRequest {
+        this['body'] = body;
+        return this;
+    }
+}