+++ services/ocr/v1/model/RecognizeHkIdCardRequest.ts
@@ -1,0 +1,12 @@
+import { HkIdCardRequestBody } from './HkIdCardRequestBody';
+
+
+export class RecognizeHkIdCardRequest {
+    public body?: HkIdCardRequestBody;
+    public constructor() { 
+    }
+    public withBody(body: HkIdCardRequestBody): RecognizeHkIdCardRequest {
+        this['body'] = body;
+        return this;
+    }
+}