+++ services/ocr/v1/model/RecognizeMainlandTravelPermitRequest.ts
@@ -1,0 +1,12 @@
+import { MainlandTravelPermitRequestBody } from './MainlandTravelPermitRequestBody';
+
+
+export class RecognizeMainlandTravelPermitRequest {
+    public body?: MainlandTravelPermitRequestBody;
+    public constructor() { 
+    }
+    public withBody(body: MainlandTravelPermitRequestBody): RecognizeMainlandTravelPermitRequest {
+        this['body'] = body;
+        return this;
+    }
+}